Update requeue_job_now.lua tests for cmd ordering
Now that LPUSH is run before ZREM, this test is invalid in its current
state. Its purpose was to ensure that trying to requeue a job that isn't
in the "source" queue already would fail. But the changes in this PR
sacrifice this guarantee to obtain a "run at least once" guarantee.